Resolve merge/conflicts.
