--- src/EPiServer.SocialAlloy.Web/Views/Social/CommentsBlock/CommentsDisplayView.cshtml
@@ -1,6 +1,0 @@
-﻿@using EPiServer.Core
-@using EPiServer.Web.Mvc.Html
-
-<div class="border">
-    Most Recent Comments Go Here
-</div>
